refactor(conversations): migrate folder out of me, delete experimental eventconnection
