+++ src/schema/v2/conversation/attachment.ts
