+++ src/schema/v2/conversation/delivery.ts
@@ -1,5 +1,5 @@
 import { GraphQLObjectType, GraphQLString, GraphQLNonNull } from "graphql"
-import date from "../../fields/date"
+import date from "schema/v2/fields/date"
 import { ResolverContext } from "types/graphql"
 import { InternalIDFields } from "schema/v2/object_identification"
