+++ src/schema/v2/conversation/index.ts
@@ -25,7 +25,6 @@ import {
   NullableIDField,
 } from "schema/v2/object_identification"
 import { MessageType } from "./message"
-import { eventConnection } from "./eventConnection"
 import { ResolverContext } from "types/graphql"
 
 export const BuyerOutcomeTypes = new GraphQLEnumType({
@@ -404,7 +403,6 @@ export const ConversationType = new GraphQLObjectType<any, ResolverContext>({
     },
 
     messagesConnection,
-    conversationEventConnection: eventConnection,
     unread: {
       type: GraphQLBoolean,
       description: "True if there is an unread message by the user.",
