+++ src/schema/v2/conversation/message.ts
