--- src/schema/v2/fields/hybridConnection/index.ts
@@ -1,1 +1,0 @@
-export { fetchHybridConnection } from "./fetchHybridConnection"
