+++ src/schema/v2/me/index.ts
@@ -29,9 +29,9 @@ import { BidderPosition } from "./bidder_position"
 import BidderPositions from "./bidder_positions"
 import BidderStatus from "./bidder_status"
 import { CollectorProfile } from "../CollectorProfile/collectorProfile"
-import Conversation from "./conversation"
-import Invoice from "./conversation/invoice"
-import Conversations from "./conversations"
+import Conversation from "schema/v2/conversation"
+import Invoice from "schema/v2/conversation/invoice"
+import Conversations from "schema/v2/conversation/conversations"
 import { CreditCards } from "./credit_cards"
 import { followedProfiles } from "./followedProfiles"
 import FollowedArtists from "./followed_artists"
