+++ src/schema/v2/object_identification.ts
@@ -49,8 +49,8 @@ const SupportedTypes: any = {
     "./home/home_page_artwork_module",
     "./home/home_page_artist_module",
     "./me",
-    "./me/conversation",
-    "./me/conversation/invoice",
+    "./conversation",
+    "./conversation/invoice",
     "./partner/partner",
     "./show",
     "./fair",
