+++ src/schema/v2/schema.ts
@@ -58,10 +58,10 @@ import { BidderPositionMutation } from "./me/bidder_position_mutation"
 // import SaleArtwork from "./sale_artwork"
 import MatchArtist from "./match/artist"
 // import { SaleArtworksConnectionField } from "./sale_artworks"
-import Conversation from "./me/conversation"
-import SendConversationMessageMutation from "./me/conversation/send_message_mutation"
-import { submitInquiryRequestMutation } from "./me/conversation/submit_inquiry_request_mutation"
-import UpdateConversationMutation from "./me/conversation/update_conversation_mutation"
+import Conversation from ".//conversation"
+import SendConversationMessageMutation from "./conversation/send_message_mutation"
+import { submitInquiryRequestMutation } from "./conversation/submit_inquiry_request_mutation"
+import UpdateConversationMutation from "./conversation/update_conversation_mutation"
 import createBidderMutation from "./me/create_bidder_mutation"
 import createCreditCardMutation from "./me/create_credit_card_mutation"
 import { deleteBankAccountMutation } from "./me/delete_bank_account_mutation"
